Moving has browser into utils.
